@@ -380,13 +380,6 @@ impl TempDir {
         self.path.as_ref()
     }
 
-    /// Deprecated alias for [`TempDir::keep`].
-    #[must_use]
-    #[deprecated = "use TempDir::keep()"]
-    pub fn into_path(self) -> PathBuf {
-        self.keep()
-    }
-
     /// Persist the temporary directory to disk, returning the [`PathBuf`] where it is located.
     ///
     /// This consumes the [`TempDir`] without deleting directory on the filesystem, meaning that
@@ -481,7 +474,10 @@ impl TempDir {
     }
 }
 
-impl AsRef<Path> for TempDir {
+// NOTE: This is implemented on &TempDir, not TempDir, to prevent accidentally moving the TempDir
+// into a function that calls `as_ref()` before immediately dropping it (deleting the underlying
+// temporary directory).
+impl AsRef<Path> for &TempDir {
     fn as_ref(&self) -> &Path {
         self.path()
     }
