@@ -42,7 +42,7 @@ pub fn create_named(
 pub fn create(dir: &Path) -> io::Result<File> {
     util::create_helper(
         dir,
-        OsStr::new(".tmp"),
+        crate::env::default_prefix(),
         OsStr::new(""),
         crate::NUM_RAND_CHARS,
         |path| {
