@@ -49,7 +49,7 @@ mod imp;
 /// # Ok::<(), std::io::Error>(())
 /// ```
 pub fn tempfile() -> io::Result<File> {
-    tempfile_in(env::temp_dir())
+    tempfile_in(env::temp_dir()?)
 }
 
 /// Create a new temporary file in the specified directory. Also see [`tempfile`].
@@ -368,13 +368,13 @@ impl Deref for TempPath {
     }
 }
 
-impl AsRef<Path> for TempPath {
+impl AsRef<Path> for &TempPath {
     fn as_ref(&self) -> &Path {
         &self.path
     }
 }
 
-impl AsRef<OsStr> for TempPath {
+impl AsRef<OsStr> for &TempPath {
     fn as_ref(&self) -> &OsStr {
         self.path.as_os_str()
     }
@@ -458,7 +458,7 @@ impl<F> fmt::Debug for NamedTempFile<F> {
     }
 }
 
-impl<F> AsRef<Path> for NamedTempFile<F> {
+impl<F> AsRef<Path> for &NamedTempFile<F> {
     #[inline]
     fn as_ref(&self) -> &Path {
         self.path()
@@ -1021,11 +1021,18 @@ impl<F: Seek> Seek for NamedTempFile<F> {
     fn seek(&mut self, pos: SeekFrom) -> io::Result<u64> {
         self.as_file_mut().seek(pos).with_err_path(|| self.path())
     }
+    fn rewind(&mut self) -> io::Result<()> {
+        self.as_file_mut().rewind().with_err_path(|| self.path())
+    }
 }
 
 impl Seek for &NamedTempFile<File> {
     fn seek(&mut self, pos: SeekFrom) -> io::Result<u64> {
         self.as_file().seek(pos).with_err_path(|| self.path())
+    }
+
+    fn rewind(&mut self) -> io::Result<()> {
+        self.as_file().rewind().with_err_path(|| self.path())
     }
 }
 
