@@ -4,7 +4,12 @@ use std::{io, iter::repeat_with};
 
 use crate::error::IoResultExt;
 
-fn tmpname(rng: &mut fastrand::Rng, prefix: &OsStr, suffix: &OsStr, rand_len: usize) -> OsString {
+fn tmpname(
+    rng: &mut fastrand::Rng,
+    prefix: &OsStr,
+    suffix: &OsStr,
+    rand_len: usize,
+) -> io::Result<OsString> {
     let capacity = prefix
         .len()
         .saturating_add(suffix.len())
@@ -16,7 +21,10 @@ fn tmpname(rng: &mut fastrand::Rng, prefix: &OsStr, suffix: &OsStr, rand_len: us
         buf.push(c.encode_utf8(&mut char_buf));
     }
     buf.push(suffix);
-    buf
+
+    check_valid_filename(&buf)?;
+
+    Ok(buf)
 }
 
 pub fn create_helper<R>(
@@ -31,8 +39,7 @@ pub fn create_helper<R>(
     let mut base = base; // re-borrow to shrink lifetime
     let base_path_storage; // slot to store the absolute path, if necessary.
     if !base.is_absolute() {
-        let cur_dir = std::env::current_dir()?;
-        base_path_storage = cur_dir.join(base);
+        base_path_storage = std::path::absolute(base)?;
         base = &base_path_storage;
     }
 
@@ -64,7 +71,7 @@ pub fn create_helper<R>(
         }
         let _ = i; // avoid unused variable warning for the above.
 
-        let path = base.join(tmpname(&mut rng, prefix, suffix, random_len));
+        let path = base.join(tmpname(&mut rng, prefix, suffix, random_len)?);
         return match f(path) {
             Err(ref e) if e.kind() == io::ErrorKind::AlreadyExists && num_retries > 1 => continue,
             // AddrInUse can happen if we're creating a UNIX domain socket and
@@ -79,4 +86,48 @@ pub fn create_helper<R>(
         "too many temporary files exist",
     ))
     .with_err_path(|| base)
+}
+
+/// Check that the passed path is a valid file-name (no directories, no nulls, etc.).
+fn check_valid_filename(path: impl AsRef<OsStr>) -> io::Result<()> {
+    let path = path.as_ref();
+
+    // Make sure the filename isn't empty.
+    if path.is_empty() {
+        return Err(io::Error::new(
+            io::ErrorKind::InvalidInput,
+            "temporary filename is empty",
+        ));
+    }
+
+    // Check for null bytes, encoding doesn't matter. The OS/libc will do this for us, but we might
+    // as well be extra safe.
+    if path.as_encoded_bytes().contains(&0) {
+        return Err(io::Error::new(
+            io::ErrorKind::InvalidInput,
+            format!("temporary filename {path:?} contains a null byte"),
+        ));
+    }
+
+    // Make sure the filename is exactly one path element and make sure that element is a file name.
+    // This is the most reliable way to check for path separators.
+    if Path::new(path).file_name() != Some(path) {
+        return Err(io::Error::new(
+            io::ErrorKind::InvalidInput,
+            format!("temporary filename {path:?} is not a valid path component"),
+        ));
+    }
+
+    Ok(())
+}
+
+#[test]
+fn test_filename_validation() {
+    check_valid_filename("foo").unwrap();
+    check_valid_filename("foo.bar").unwrap();
+    check_valid_filename("/foo").expect_err("path contains a slash");
+    check_valid_filename("foo/bar").expect_err("path contains a slash");
+    check_valid_filename("foo/").expect_err("path contains a slash");
+    check_valid_filename("/").expect_err("path contains a slash");
+    check_valid_filename("\0").expect_err("path contains a null");
 }
